pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DDMethods"
version = "0.1.0"
description = "Package for Data Driven Methods Engineering"
readme = "README.md"
requires-python = "==3.13.*"
license = { text = "MIT" }
authors = [
    { name = "Philip Korsager Nickel", email = "s214960@dtu.dk" }
]

# All dependencies (runtime, docs, and dev)
dependencies = [
  "hydra-core>=1.3.2",
  "mlflow>=3.9.0",
  # Runtime dependencies
  "torch>=2.10.0",
  "torchvision>=0.25.0",
  # Scientific computing
  "numpy>=1.26.0",
  "scipy>=1.14.0",
  "matplotlib>=3.9.0",
  "wandb>=0.25.0",
  "dtumathtools>=2025.2.0",
  "ipykernel>=7.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 99
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["ML"]

